feat: 3주차 코드 구현

+++ src/main/java/codestartup/codestartup/order/domain/repository/BookRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface BookRepository {
     List<Book> findAll();
-    Optional<Book> findById(Integer id);
+    Optional<Book> findById(Long id);
 }
